backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
django-filter==23.5
Pillow==9.3.0
djoser==2.1.0
django-urlshortner==0.0.2
gunicorn==20.1.0
psycopg2-binary==2.9.3 